Add user and group endpoints and user fetching
Begin to implement user service endpoints, starting with the
`/user/me/` and `/groups` to get user info back to the client
and sort out paginated results. This also implements the logic
in the client to refresh the Google OAuth2 access token and
recreate the Yonder token that is passed around to auth other
Yonder services.